Remove Green River sources in root
