+++ green-river/src/main/scala/consumer/activity/CouponConnector.scala
@@ -6,20 +6,23 @@ import consumer.aliases._
 
 import org.json4s.JsonAST.{JInt, JString, JNothing}
 
-object CouponConnector extends ActivityConnector {
+final case class CouponConnector()(implicit ec: EC) extends ActivityConnector {
   val dimension = "coupon"
 
-  def process(offset: Long, activity: Activity)(implicit ec: EC): Future[Seq[Connection]] =
-    Future {
-      byNoteData(activity).distinct.map(createConnection(_, activity.id))
-    }
+  def process(offset: Long, activity: Activity): Future[Seq[Connection]] = Future {
+    val couponIds = byNoteData(activity)
 
-  def createConnection(formId: String, activityId: Int): Connection =
+    couponIds.distinct.map(createConnection(_, activity.id))
+  }
+
+  def createConnection(formId: String, activityId: Int): Connection = {
     Connection(dimension = dimension, objectId = formId, data = JNothing, activityId = activityId)
+  }
 
-  private def byNoteData(activity: Activity): Seq[String] =
+  private def byNoteData(activity: Activity): Seq[String] = {
     (activity.data \ "note" \ "referenceType", activity.data \ "entity" \ "id") match {
       case (JString(`dimension`), JInt(id)) ⇒ Seq(id.toString)
       case _                                ⇒ Seq.empty
     }
+  }
 }
